Fix github_issue_id overflow — widen INTEGER to BIGINT
GitHub issue IDs (e.g. 4594515575) exceed PostgreSQL's 32-bit INTEGER
max of ~2.1B. Change all three github_issue_id columns to BIGINT in the
model and add an ALTER COLUMN migration that runs on startup.

+++ app/models.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime
-from sqlalchemy import Column, String, Integer, DateTime, Text, Boolean, JSON, ForeignKey
+from sqlalchemy import Column, String, Integer, BigInteger, DateTime, Text, Boolean, JSON, ForeignKey
 from sqlalchemy.orm import relationship
 from .database import Base
 
@@ -124,7 +124,7 @@ class Project(Base):
     github_issue_url = Column(String, nullable=True)
     github_issue_number = Column(Integer, nullable=True)
     github_issue_node_id = Column(String, nullable=True)
-    github_issue_id = Column(Integer, nullable=True)  # numeric REST id (sub_issue_id)
+    github_issue_id = Column(BigInteger, nullable=True)  # numeric REST id (sub_issue_id)
     github_project_item_id = Column(String, nullable=True)
     # Per-project GitHub Project board (Option B — one board per project)
     github_project_node_id = Column(String, nullable=True)   # board GraphQL ID
@@ -158,7 +158,7 @@ class MeetingNote(Base):
     github_issue_url = Column(String, nullable=True)
     github_issue_number = Column(Integer, nullable=True)
     github_issue_node_id = Column(String, nullable=True)
-    github_issue_id = Column(Integer, nullable=True)  # numeric REST id (sub_issue_id)
+    github_issue_id = Column(BigInteger, nullable=True)  # numeric REST id (sub_issue_id)
     github_project_item_id = Column(String, nullable=True)
     reviewer_github_username = Column(String, nullable=True)
     reviewer_name = Column(String, nullable=True)
@@ -234,7 +234,7 @@ class PrdDocument(Base):
     github_issue_url = Column(String, nullable=True)
     github_issue_number = Column(Integer, nullable=True)
     github_issue_node_id = Column(String, nullable=True)
-    github_issue_id = Column(Integer, nullable=True)  # numeric REST id (sub_issue_id)
+    github_issue_id = Column(BigInteger, nullable=True)  # numeric REST id (sub_issue_id)
     github_project_item_id = Column(String, nullable=True)
     reviewer_github_username = Column(String, nullable=True)
     reviewer_name = Column(String, nullable=True)
